tTest mit added Library

+++ src/src/calculateRowStatistics.java
@@ -32,9 +32,9 @@ public class calculateRowStatistics {
         return group;
     }
 
-    public static List<List<Integer>> meanCalculator(Counts counts){
+    public static List<List<Double>> meanCalculator(Counts counts){
         List<List<Integer>> count = counts.getCounts();
-        List<List<Integer>> meanValues = new ArrayList<>();
+        List<List<Double>> meanValues = new ArrayList<>();
         int[] group = groupDefiner(counts);
         int rownumber = count.size();
         int colnumber = count.get(0).size();
@@ -54,9 +54,9 @@ public class calculateRowStatistics {
                         MagnitudeTwo = MagnitudeTwo + row.get(j);
                     }
                 }
-                int meanOne = (CounterOne != 0) ? MagnitudeOne / CounterOne : 0;
-                int meanTwo = (CounterTwo != 0) ? MagnitudeTwo / CounterTwo : 0;
-                List<Integer> groupMeans = new ArrayList<>();
+                double meanOne = (CounterOne != 0) ? MagnitudeOne / CounterOne : 0;
+                double meanTwo = (CounterTwo != 0) ? MagnitudeTwo / CounterTwo : 0;
+                List<Double> groupMeans = new ArrayList<>();
                 groupMeans.add(meanOne);
                 groupMeans.add(meanTwo);
                 meanValues.add(groupMeans);
